10주차 > basic > ScheduleTables 컴포넌트 분리, dialog open state 설정, api 분리
