+++ packages/assignment-10/src/basic/constants/index.ts
@@ -1,3 +1,5 @@
+import { parseHnM } from "../utils";
+
 export const DAY_LABELS = ["월", "화", "수", "목", "금", "토"] as const;
 
 export const CellSize = {
@@ -7,6 +9,18 @@ export const CellSize = {
 
 export const 초 = 1000;
 export const 분 = 60 * 초;
+
+export const TIMES = [
+  ...Array(18)
+    .fill(0)
+    .map((v, k) => v + k * 30 * 분)
+    .map((v) => `${parseHnM(v)}~${parseHnM(v + 30 * 분)}`),
+
+  ...Array(6)
+    .fill(18 * 30 * 분)
+    .map((v, k) => v + k * 55 * 분)
+    .map((v) => `${parseHnM(v)}~${parseHnM(v + 50 * 분)}`),
+] as const;
 
 export const TIME_SLOTS = [
   { id: 1, label: "09:00~09:30" },
