+++ packages/assignment-10/src/basic/types/index.ts
@@ -28,3 +28,8 @@ export interface SearchOption {
   majors: string[];
   credits?: number;
 }
+
+export interface TimeInfo {
+  day: string;
+  time: number;
+}
